Move most of index to ski-free, allows functions to be more testable
